Split off client entity load events into new class
