+++ library/entity/entity-events/src/main/java/org/quiltmc/qsl/entity/mixin/ClientEntityHandlerMixin.java
@@ -18,7 +18,7 @@ package org.quiltmc.qsl.entity.mixin;
 
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.entity.Entity;
-import org.quiltmc.qsl.entity.api.event.EntityLoadEvents;
+import org.quiltmc.qsl.entity.api.event.client.ClientEntityLoadEvents;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -32,11 +32,11 @@ public abstract class ClientEntityHandlerMixin {
 
 	@Inject(method = "startTracking(Lnet/minecraft/entity/Entity;)V", at = @At("TAIL"))
 	void invokeEntityLoadEvent(Entity entity, CallbackInfo ci) {
-		EntityLoadEvents.AFTER_ENTITY_LOAD_CLIENT.invoker().onLoad(entity, this.field_27735);
+		ClientEntityLoadEvents.AFTER_ENTITY_LOAD_CLIENT.invoker().onLoad(entity, this.field_27735);
 	}
 
 	@Inject(method = "stopTracking(Lnet/minecraft/entity/Entity;)V", at = @At("TAIL"))
 	void invokeEntityUnloadEvent(Entity entity, CallbackInfo ci) {
-		EntityLoadEvents.AFTER_ENTITY_UNLOAD_CLIENT.invoker().onUnload(entity, this.field_27735);
+		ClientEntityLoadEvents.AFTER_ENTITY_UNLOAD_CLIENT.invoker().onUnload(entity, this.field_27735);
 	}
 }
